Validade lock across multiple clients
Add the possibility of checking if an account lock is valid across
multiple redis instances. Following some structures detailed on the
Redlock algorithm, we're able to validate an account lock even on a
multiple Redis infrastructure.

A basic log structure was also added to support the validation part of
this feature. The intention here is to expand the use of this logger
throughout the application.