Add retries for when locks fail
When trying to lock an account it's imperative to consider concurrency,
for this reason we're implementing a way to retry when creating this lock.

The default parameters are: 3 retries, 1 second apart from each other.
We are considering that 1 second is more than enough for the process of
creating a lock, registering the movements and releasing the said lock.
In a environment with multiple movements begin created at the same time,
we're leaving a retry count of 3, meaning that only after 3 seconds that
request would be marked as invalid and the client should deal with it.